requirements.txt
autopep8
cffi
click
cryptography
Flask
gunicorn
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
psycopg2
pycodestyle
pycparser
python-dotenv
toml
Werkzeug
zipp
cryptography
flask-migrate
SQLAlchemy